# gen/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "contact": {
            "email": "support@codespace.dev",
            "name": "Codespace Operator Support",
            "url": "https://github.com/codespace-operator/codespace-operator"
        },
        "description": "REST API for managing Codespace sessions with RBAC and multi-provider authentication",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "termsOfService": "https://github.com/codespace-operator/codespace-operator/blob/main/TERMS.md",
        "title": "Codespace Operator API",
        "version": "2.0.0"
    },
    "paths": {
        "/api/v1/admin/rbac/reload": {
            "post": {
                "description": "Force reload of RBAC policies (requires admin privileges)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "Reload RBAC (Admin)",
                "tags": [
                    "admin"
                ]
            }
        },
...
        "/api/v1/admin/system/info": {
            "get": {
                "description": "Check system information (requires admin privileges)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "System Info",
                "tags": [
                    "admin"
                ]
            }
        },
        "/api/v1/admin/users": {
            "get": {
                "description": "Get list of users in the system (requires admin privileges)",
                "responses": {
                    "200": {
                        "description": "User list with admin info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "List users (Admin)",
                "tags": [
                    "admin"
                ]
            }
        },
        "/api/v1/introspect": {
            "get": {
                "deprecated": true,
                "description": "Deprecated: prefer /api/v1/introspect/user or /api/v1/introspect/server",
                "responses": {
                    "200": {
                        "description": "Combined user+server info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "Introspect (legacy combined)",
                "tags": [
                    "user"
                ]
            }
        },
        "/api/v1/introspect/server": {
            "get": {
                "description": "Get server and cluster information",
                "parameters": [
                    {
                        "description": "Whether to discover namespaces (0 or 1)",
                        "in": "query",
                        "name": "discover",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ServerIntrospectionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "Server introspection",
                "tags": [
                    "user"
                ]
            }
        },
        "/api/v1/introspect/user": {
            "get": {
                "description": "Get user-specific permissions and capabilities",
                "parameters": [
                    {
                        "description": "Comma-separated list of namespaces to check",
                        "in": "query",
                        "name": "namespaces",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Comma-separated list of actions to check",
                        "in": "query",
                        "name": "actions",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.UserIntrospectionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "User introspection",
                "tags": [
                    "user"
                ]
            }
        },
        "/api/v1/me": {
            "get": {
                "description": "Get information about the current authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.UserInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "Get current user",
                "tags": [
                    "user"
                ]
            }
        },
...
        "/api/v1/stream/sessions": {
            "get": {
                "description": "Stream real-time session updates via Server-Sent Events",
                "operationId": "streamSessions",
                "parameters": [
                    {
                        "description": "Target namespace",
                        "in": "query",
                        "name": "namespace",
                        "schema": {
                            "type": "string",
                            "default": "default"
                        }
                    },
                    {
                        "description": "Stream sessions from all namespaces",
                        "in": "query",
                        "name": "all",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream",
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "Stream sessions",
                "tags": [
                    "sessions"
                ]
            }
        },
...
    },
...
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer token authentication (JWT)",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            },
            "CookieAuth": {
                "description": "Session cookie authentication",
                "in": "cookie",
                "name": "codespace_session",
                "type": "apiKey"
            }
        },
        "schemas": {
...
            "internal_server.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "Invalid request",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "internal_server.ServerIntrospectionResponse": {
                "properties": {
                    "capabilities": {
                        "$ref": "#/components/schemas/internal_server.SystemCapabilities"
                    },
                    "cluster": {
                        "$ref": "#/components/schemas/internal_server.ClusterInfo"
                    },
                    "instanceID": {
                        "type": "string"
                    },
                    "manager": {
                        "$ref": "#/components/schemas/common.AnchorMeta"
                    },
                    "namespaces": {
                        "$ref": "#/components/schemas/internal_server.ServerNamespaceInfo"
                    },
                    "version": {
                        "$ref": "#/components/schemas/internal_server.ServerVersionInfo"
                    }
                },
                "type": "object"
            },
...
            "internal_server.UserIntrospectionResponse": {
                "properties": {
                    "capabilities": {
                        "$ref": "#/components/schemas/internal_server.UserCapabilities"
                    },
                    "domains": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/internal_server.DomainPermissions"
                        },
                        "type": "object"
                    },
                    "namespaces": {
                        "$ref": "#/components/schemas/internal_server.NamespaceInfo"
                    },
                    "user": {
                        "$ref": "#/components/schemas/internal_server.UserInfo"
                    }
                },
                "type": "object"
            },
...
        }
    }
}